+++ src/NLog/Internal/ThreadIDHelper.cs
@@ -47,7 +47,7 @@ namespace NLog.Internal
         /// </summary>
         static ThreadIDHelper()
         {
-#if !SILVERLIGHT && !__IOS__ && !__ANDROID__
+#if !SILVERLIGHT && !__IOS__ && !__ANDROID__ && !NETSTANDARD
             if (PlatformDetector.IsWin32)
             {
                 Instance = new Win32ThreadIDHelper();