+++ src/NLog/Layouts/JsonLayout.cs
@@ -86,7 +86,7 @@ namespace NLog.Layouts
         /// </summary>
         public bool IncludeMdc { get; set; }
 
-#if NET4_0 || NET4_5
+#if !SILVERLIGHT
         /// <summary>
         /// Gets or sets a value indicating whether to include contents of the <see cref="MappedDiagnosticsLogicalContext"/> dictionary.
         /// </summary>
@@ -117,7 +117,7 @@ namespace NLog.Layouts
             {
                 base.ThreadAgnostic = false;
             }
-#if NET4_0 || NET4_5
+#if !SILVERLIGHT
             if (IncludeMdlc)
             {
                 base.ThreadAgnostic = false;
@@ -183,7 +183,7 @@ namespace NLog.Layouts
                 }
             }
 
-#if NET4_0 || NET4_5
+#if !SILVERLIGHT
             if (this.IncludeMdlc)
             {
                 foreach (string key in MappedDiagnosticsLogicalContext.GetNames())