+++ tests/NLog.UnitTests/Config/InternalLoggingTests.cs
@@ -63,7 +63,7 @@ namespace NLog.UnitTests.Config
         [Fact]
         public void InternalLoggingConfigTestDefaults()
         {
-            using (new InternalLoggerScope())
+            using (new InternalLoggerScope(true))
             {
                 InternalLogger.LogLevel = LogLevel.Error;
                 InternalLogger.LogToConsole = true;
@@ -71,9 +71,7 @@ namespace NLog.UnitTests.Config
                 LogManager.GlobalThreshold = LogLevel.Fatal;
                 LogManager.ThrowExceptions = true;
                 LogManager.ThrowConfigExceptions = null;
-#if !__IOS__ && !__ANDROID__
                 InternalLogger.LogToTrace = true;
-#endif
 
                 CreateConfigurationFromString(@"
 <nlog>
@@ -85,38 +83,39 @@ namespace NLog.UnitTests.Config
                 Assert.Same(LogLevel.Fatal, LogManager.GlobalThreshold);
                 Assert.True(LogManager.ThrowExceptions);
                 Assert.Null(LogManager.ThrowConfigExceptions);
-#if !__IOS__ && !__ANDROID__
                 Assert.True(InternalLogger.LogToTrace);
-#endif
             }
         }
 
         [Fact]
         public void InternalLoggingConfig_off_should_be_off()
         {
-            var sb = new StringBuilder();
-            var stringWriter = new StringWriter(sb);
-            InternalLogger.LogWriter = stringWriter;
-            string wrongFileName = "WRONG/***[]???////WRONG";
-            LogManager.Configuration = CreateConfigurationFromString(string.Format(@"<?xml version='1.0' encoding='utf-8' ?>
-<nlog internalLogFile='{0}'
-      internalLogLevel='Off'
-      throwExceptions='true' >
+            using (new InternalLoggerScope())
+            {
+                var sb = new StringBuilder();
+                var stringWriter = new StringWriter(sb);
+                InternalLogger.LogWriter = stringWriter;
+                string wrongFileName = "WRONG/***[]???////WRONG";
+                LogManager.Configuration = CreateConfigurationFromString(string.Format(@"<?xml version='1.0' encoding='utf-8' ?>
+    <nlog internalLogFile='{0}'
+          internalLogLevel='Off'
+          throwExceptions='true' >
 
-  <targets>
-    <target name='logfile' type='File' fileName='WRONG'  />
-  </targets>
+      <targets>
+        <target name='logfile' type='File' fileName='WRONG'  />
+      </targets>
 
-  <rules>
-    <logger name='*' writeTo='logfile' />
-  </rules>
-</nlog>
-", wrongFileName));
+      <rules>
+        <logger name='*' writeTo='logfile' />
+      </rules>
+    </nlog>
+    ", wrongFileName));
 
-            Assert.Equal("",sb.ToString());
-            Assert.Equal(LogLevel.Off,InternalLogger.LogLevel);
-            Assert.False(InternalLogger.ExceptionThrowWhenWriting);
+                Assert.Equal("", sb.ToString());
+                Assert.Equal(LogLevel.Off, InternalLogger.LogLevel);
+                Assert.False(InternalLogger.ExceptionThrowWhenWriting);
             }
+        }
 
         private void InternalLoggingConfigTest(LogLevel logLevel, bool logToConsole, bool logToConsoleError, LogLevel globalThreshold, bool throwExceptions, bool? throwConfigExceptions, string file, bool logToTrace)
         {
@@ -128,7 +127,7 @@ namespace NLog.UnitTests.Config
             var throwConfigExceptionsString = throwConfigExceptions == null ? "" : throwConfigExceptions.ToString().ToLower();
             var logToTraceString = logToTrace.ToString().ToLower();
 
-            using (new InternalLoggerScope())
+            using (new InternalLoggerScope(true))
             {
                 CreateConfigurationFromString(string.Format(@"
 <nlog internalLogFile='{0}' internalLogLevel='{1}' internalLogToConsole='{2}' internalLogToConsoleError='{3}' globalThreshold='{4}' throwExceptions='{5}' throwConfigExceptions='{6}' internalLogToTrace='{7}'>
@@ -148,9 +147,7 @@ namespace NLog.UnitTests.Config
 
                 Assert.Equal(throwConfigExceptions, LogManager.ThrowConfigExceptions);
 
-#if !__IOS__ && !__ANDROID__
                 Assert.Equal(logToTrace, InternalLogger.LogToTrace);
-#endif
             }
         }
     }