+++ tests/NLog.UnitTests/Internal/LayoutHelpersTests.cs
@@ -31,11 +31,6 @@
 // THE POSSIBILITY OF SUCH DAMAGE.
 // 
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using NLog.Internal;
 using NLog.Layouts;
 using Xunit;