+++ tests/NLog.UnitTests/LayoutRenderers/CallSiteLineNumberTests.cs
@@ -49,10 +49,8 @@ namespace NLog.UnitTests.LayoutRenderers
 
     public class CallSiteLineNumberTests : NLogTestBase
     {
-
-
-#if MONO
-        [Fact(Skip="Not working under MONO - not sure if unit test is wrong, or the code")]
+#if !DEBUG
+        [Fact(Skip = "RELEASE not working, only DEBUG")]
 #else
         [Fact]
 #endif