+++ CHANGELOG.md
@@ -4,7 +4,7 @@ See also [releases](https://github.com/NLog/NLog/releases) and [milestones](http
 
 #### Fixes
 
-- [#2229] Fix: ReconfigExistingLoggers sometimes throws an Exception (@jpdillingham)
+- [#2229](https://github.com/nlog/nlog/pull/2229) Fix: ReconfigExistingLoggers sometimes throws an Exception (@jpdillingham)
 
 ### v4.4.11 (2017/06/17)